refactor: move architecture to MVC pattern
